chore: move parseDecimals function to utils
