+++ examples/oft-solana/tasks/solana/utils.ts
@@ -75,3 +75,18 @@ export const findSolanaEndpointIdInGraph = async (
 
     throw new Error('No Solana Endpoint ID found. Ensure your OApp configuration includes a valid Solana endpoint.')
 }
+
+/**
+ * Turn a human decimal amount (e.g. "1.234") into a BigInt of base‐units given `decimals`.
+ */
+export function parseDecimalToUnits(amount: string, decimals: number): bigint {
+    const [whole, fraction = ''] = amount.split('.')
+    const wholeUnits = BigInt(whole) * 10n ** BigInt(decimals)
+    // pad or trim the fractional part to exactly `decimals` digits
+    const fracUnits = BigInt(
+        fraction
+            .padEnd(decimals, '0') // "23"  → "230000"
+            .slice(0, decimals) // in case user typed more digits than `decimals`
+    )
+    return wholeUnits + fracUnits
+}
